Refactor tests around waiting and stopping
