+++ docs/conf.py
@@ -14,6 +14,7 @@
 # import sys
 # sys.path.insert(0, os.path.abspath('.'))
 
+from typing import List
 
 # -- Project information -----------------------------------------------------
 
@@ -49,6 +50,6 @@ html_theme = "alabaster"
 # relative to this directory. They are copied after the builtin static files,
 # so a file named "default.css" will overwrite the builtin "default.css".
 # html_static_path = ["_static"]
-html_static_path = []
+html_static_path: List[str] = []
 
 autoclass_content = "both"
